refactor: moved gamePlayer gui to its own directory
Issue #73
