+++ src/gui/__init__.py
@@ -1,3 +1,2 @@
 from .gameCreationPage import GameCreationPage
 from .homePage import HomePage
-from .settingsPage import SettingsPage
